### 2018-01-16 #534 @arfon
@acabunoc - would you be willing to edit this submission for JOSS?
