### 2018-01-26 #534 @acabunoc
What a showing! Thanks @paulklemm & @dvalters 😄
Here are the reviewer guidelines: http://joss.theoj.org/about#reviewer_guidelines

I'll take a quick look at your paper.md before we start the review
