### 2018-01-26 #534 @acabunoc
@jorgeboucas could you merge [this PR](https://github.com/mpg-age-bioinformatics/bit/pull/5)? I _think_ it will fix the error @whedon's running into
