### 2018-01-26 #534 @jorgeboucas
@acabunoc merged. Thanks!
