### 2018-01-26 #534 @jorgeboucas
@acabunoc `paper.md` functional now.
