### 2018-01-26 #549 @whedon
[REVIEW]: bit: a git-based tool for the management of code and data
**Submitting author:** @jorgeboucas (<a href="http://orcid.org/0000-0002-9412-7592">Jorge Boucas</a>)
**Repository:** <a href="https://github.com/mpg-age-bioinformatics/bit" target ="_blank">https://github.com/mpg-age-bioinformatics/bit</a>
**Version:** 0.1.1
