### 2018-02-13 mpg-age-bioinformatics/bit#6 @dvalters
Traceback running bit --config
Discovered through the JOSS review process over at: https://github.com/openjournals/joss-reviews/issues/549#

Downloaded and run setup.py as per the instructions in the README, but on running `bit --config` in the first example given I get Traceback:

(The error is the same for `bit --start`, i.e. `TabError`)
```
$ bit --config
Traceback (most recent call last):
  File "/home/dav/.local/bin/bit", line 11, in <module>
    load_entry_point('bit==0.1.0', 'console_scripts', 'bit')()
  File "/home/dav/anaconda3/lib/python3.5/site-packages/setuptools-27.2.0-py3.5.egg/pkg_resources/__init__.py", line 565, in load_entry_point
  File "/home/dav/anaconda3/lib/python3.5/site-packages/setuptools-27.2.0-py3.5.egg/pkg_resources/__init__.py", line 2598, in load_entry_point
  File "/home/dav/anaconda3/lib/python3.5/site-packages/setuptools-27.2.0-py3.5.egg/pkg_resources/__init__.py", line 2258, in load
  File "/home/dav/anaconda3/lib/python3.5/site-packages/setuptools-27.2.0-py3.5.egg/pkg_resources/__init__.py", line 2264, in resolve
  File "/home/dav/.local/lib/python3.5/site-packages/bit-0.1.0-py3.5.egg/bit/__init__.py", line 11, in <module>
    import bit.config as config
  File "/home/dav/.local/lib/python3.5/site-packages/bit-0.1.0-py3.5.egg/bit/config.py", line 112
    return user_group
                    ^
TabError: inconsistent use of tabs and spaces in indentation
```
