### 2018-02-13 mpg-age-bioinformatics/bit#7 @dvalters
Feedback on JOSS paper (paper.md)
(I've put paper issues in this issue tracker rather than leave comments on the review page as per the JOSS guidelines https://github.com/openjournals/joss-reviews/issues/549)

[Line numbers refer to line breaks when opened in text editor]

## Suggested changes for style/grammar

**l.21**

> In the increasing world of data

Change to something like: _"With the increasing amount of data in the world..."_

> the mingles of data

Change to _"mingling of data"_

**l.23**

> proficient tools have emerged along the last years

Change to: _"emerged in recent years"_

> a unified tool is still in the need for many.

Change to something like: _"...is still **a** need for many [researchers]"_

**l.24** - not quite clear what this means - could do with rewording.

**l.26**
remove first "for example"
`eg.` -> `e.g.`

**l.28**

> Project metadata and data can like this be kept in different instances.

Remove _"like this"_ or reword to _"can be kept like this in different instances"_

**l.30**

> Using rsync, it allows seamless multiple rsync calls for efficient synchronisations

Suggested change to _"~Using rsync~ bit allows seamless multiple rsync calls..."_.

**l.32**

Add punctuation somewhere to break long sentence, suggest comma after _"HPC clusters,"_

## References/bibliography

There are no references/citations in the paper and the `.bib` file is blank (on master - it is not present in the tagged release version). Check that there are no references that can be added. (I assume this is OK if there is actually nothing to be cited, but maybe check with the JOSS editors)
