### 2018-02-13 #549 @dvalters
Hi @jorgeboucas, I've raised a few issues in the issue tracker in the submission's github repository (see linked issues above). My other comments so far are:

1. Documentation is a bit sparse - there is a 'man page' style usage guide in the README.md file, which I think might meet the minimum requirements for documentation (@acabunoc?) but I could not find any separate documentation linked for the software. (My impression is that JOSS submissions should have separate documentation pages _in addition_ to a high-level overview in the README file, but please correct me if I'm wrong @acabunoc)

2. No tests per se. There are examples of usage in the README.md file, but again, these are fairly limited, and I'm not sure they are sufficient under the JOSS guidelines:

> Automated tests: Are there automated tests or manual steps described so that the function of the software can be verified?

3. No community guidelines present.

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
