### 2018-03-19 #549 @arfon
@paulklemm - please carry out your review here when you get a chance?

@jorgeboucas - also, please take a look at @dvalters' comments and provide feedback.
