### 2018-06-15 #549 @dvalters
@arfon ready to proceed, but https://github.com/mpg-age-bioinformatics/bit/issues/6 question needs resolved before can test software
