### 2018-07-03 mpg-age-bioinformatics/bit#8 @paulklemm
Add more detailed instructions on initial setup
Issue as part of JOSS review process (https://github.com/openjournals/joss-reviews/issues/549)

The `bit --config` part of the *Example usage* section worked fine for me.

After this step I feel there should be some more information available on which folders need to be set up in owncloud and how the folder structure should look on the HPC. I cannot make the `bit --start` work as I get the following output with nothing happening:

```
The bermuda information triangle works on the basis that all your projects are located in the same path and have a parent subpath in your local machine ie.
/file_system_a
    |
    '- data
        |
        '- projects
            |
            |- Company_A
            |   |
            |   |- CA_project_y
            |   |
            |   '- CA_project_x
            |       |
            |       |- results
            |       |- models
            |       |- scripts
            |       |- tmp
            |       |- slurm_logs
            |       '- wiki
            |
            '- Company_B
                |
                '- CB_project_n
```

Please expand the documentation here or allow bit to scaffold of a example structure.
