### 2018-07-03 mpg-age-bioinformatics/bit#9 @paulklemm
Suggestion: Use Tokens for Github Auhentication
Issue as part of JOSS review process (openjournals/joss-reviews#549).

`bit` currently uses username and password for github.com or enterprise github authentication. This will likely fail for me because I've enabled two-factor authentication for my accounts. `bit` should switch to personal access tokens which is also safer and allows to tweak permissions.
