### 2018-07-03 #549 @paulklemm
Also ready to proceed, require issue https://github.com/mpg-age-bioinformatics/bit/issues/8 to be solved to test the software.
