### 2018-10-04 mpg-age-bioinformatics/bit#9 @jorgeboucas
you can also `--gitssh`
