### 2018-10-04 04f8dd5
change to default use of ssh for git ; added information on tokens where git password is requested ; closes #9
Co-authored-by: drosskopp <info@upcal.de>
