### 2018-11-02 #549 @arfon
Hi @jorgeboucas - are you still interested in pursuing this publication in JOSS?
