### 2018-11-07 ddbd597
git dev has now all required functions ready to be deployed
